+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `libherokubuildpack`:
+  - Added `inventory::schedule` module with `Schedule` and `Release` types for tracking version lifecycle data such as end-of-life dates. ([#989](https://github.com/heroku/libcnb.rs/pull/989))
 
 ## [0.30.3] - 2026-03-02