+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `libcnb-test`:
+  - Added `assert_matches!` macro for asserting that expressions match specific patterns, with optional guards. ([#996](https://github.com/heroku/libcnb.rs/pull/996))
+
 
 ## [0.30.3] - 2026-03-02