+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.30.4] - 2026-04-13
+
 ### Added
 
 - `libcnb-test`:
@@ -450,7 +453,8 @@ version number. See the changelog below for other changes.
 
 - Remove support for legacy BOM. Remove `Launch::bom`, `Build::bom`, `bom::Bom`, `bom::Entry`. ([#489](https://github.com/heroku/libcnb.rs/pull/489))
 
-[unreleased]: https://github.com/heroku/libcnb.rs/compare/v0.30.3...HEAD
+[unreleased]: https://github.com/heroku/libcnb.rs/compare/v0.30.4...HEAD
+[0.30.4]: https://github.com/heroku/libcnb.rs/compare/v0.30.3...v0.30.4
 [0.30.3]: https://github.com/heroku/libcnb.rs/compare/v0.30.2...v0.30.3
 [0.30.2]: https://github.com/heroku/libcnb.rs/compare/v0.30.1...v0.30.2
 [0.30.1]: https://github.com/heroku/libcnb.rs/compare/v0.30.0...v0.30.1